Migration of brain to server compatible version
