+++ core/__init__.py
@@ -2,8 +2,7 @@ from .memory import Async_DB_Interface
 from .memory.weaviate_db.weaviate_db import Weaviate
 from .memory.weaviate_db.weaviate_handler import WeaviateHelper
 from .memory.weaviate_db.weaviate_utils import *
-# from .brain.main import Brain, Model
-from .brain.main_server import Brain, Model
+from .brain.main import Brain, Model
 from .event_manager.async_eda import PubSub
 from .brain.brain_handler import BrainHelper
